Add models for surveys.
