+++ home/models/__init__.py
@@ -1,0 +1,4 @@
+from .blog import *
+from .event import *
+from .session import *
+from .survey import *
